Added 31 similar IR nodes to the IR Fuzzer.
I added more IR nodes, specifically IR ops similar to Add because we already have infrastructure to easily add these nodes. This required updating the proto, visitor, and IR node generator. A lot of testing and debugging was done to iron out test cases associated with different ops. This CL is large because most of the code dealing with Ops is repetative and configuring large protobuf tests.

PiperOrigin-RevId: 781251643