Added arrays and tuples to the IR Fuzzer.
Arrays and tuples are recursive types in the sense that they contain sub-types. In order to coerce between recursive types, we must recursively iterate over the sub-types and coerce them into the correct type. This is currently done by using a default value if the types mismatch. Same categorical type conversion for tuples and arrays simply extends or shrinks the list with default types, which is done by recreating the list type.

PiperOrigin-RevId: 781601034