Control opt pipeline using toolchains.
Toolchains allow one to configure aspects of compilation in a reasonably straightforward way. This change makes a toolchain responsible for selecting the default pass pipeline used by XLS.

Note that this does not currently allow toolchains to control the actual tools used in compilation and the separate (confusingly named) xls_toolchain remains the only way to do that and is unchanged.

PiperOrigin-RevId: 781840644