Allow passing a custom optimization_pipeline proto to opt_main
When experimenting with alternative optimization pipelines it is useful to be able to pass it manually to opt_main and examine the results. This adds a --pipeline_proto and --pipeline_textproto flag which allow one to replace the default pipeline and any compound passes with the ones defined in that proto.

PiperOrigin-RevId: 782083029