Added 18 more complex bits type Ops to the IR Fuzzer.
Added more complex Ops that have the option of dealing with only the bits type were added, presumably all of the rest of them. For example, Match and OneHot. This CL does not support non-bits types, so any Ops that have the ability to use other types is not functional. This CL is large because most of the code dealing with Ops is repetative and configuring large protobuf tests.

PiperOrigin-RevId: 782123196