Token simplification pass runtime optimization.
Perhaps surprisingly, this was showing up in profiles as a significant contributor to runtime. We avoid building up sets of all transitive dependencies and pairwise-checking operands and instead analyze operands of after_alls. For each node that is an operand of an after_all, we check if any of its transitive users are also an operand of one of those after_alls. This by itself does not improve runtime, but an additional optimization that uses the already-constructed topo-sort to prune transitive users that occur after the after_all in the topo-sort gives a significant speedup. For whatever reason, a similar optimization to the previous set-of-transitive-operands approach did not yield a similar speedup. It's possible this is overfit to a specific design.

PiperOrigin-RevId: 782146044