Let ProcElaboration calculate the channel graph for a proc network.
It's often useful to be able to query the channel connection topology of a proc instantiation. This adds that ability to the ProcElaboration.

With `ProcElaboration::GetChannelGraph()` one gets an or-tools graph where nodes are each proc instantiation and edges are channels from 'sender -> receiver'. The ProcElaboration has accessors to transform the ortools node and arc ids back into proc-instantiations and send/recv pairs.

PiperOrigin-RevId: 784684316