Refactored the IR Fuzzer to prepare for adding types; includes coercion associated with types.
The protobuf information that is associated with bit width fitting, which is a method of coercing between categorical bits types, is now generated along with the type. Coercing operands now uses this CoercedType because it makes more sense for a type to specify how an operand is coerced into it.

Multiple type protobufs are used which are traversed in a recursive templated format to retrieve information from them. Each type protobuf accomplishes a different task. They need to be separate for enforced randomization structure purposes.

PiperOrigin-RevId: 785613861