[codegen] Only send active outputs when using channel multiplexing
Previously, if stage A contained a single predicated `send` for channel C, its predicate was ignored by the logic that computed the data output. This could cause issues in cases where the `send` in stage A was disabled, but another `send` on channel C in some other stage B was enabled; stage A's disabled `send` output would end up OR'd into the data that was supposed to be entirely controlled by stage B.

We now guarantee that if a `send` is disabled, it will not contribute to the data line at any time when another `send` might be enabled.

PiperOrigin-RevId: 785622722