Make ContextRangeQueryEngine handle contradictions more gracefully.
Previously we would continue trying to feedback given information even if the contextual information contains a contradiction (eg saying the selector in `(select 0b0 X Y)` is 0b1). This could lead to unexpected results such as the query engine saying that literals have different values then their actual one.

Ideally we'd communicate that there are contradictions in the set of context's were in since this information could be used for optimization. For now however simply fallback to the default range analysis whenever a contradiction is found.

This bug was found by the ir-fuzzer on narrowing-pass. This also adds a fuzz test for the context query engine to hopefully avoid this happening in the future.

Unfortunately bug 2674 can still occur since the context qe may not have enough information to realize the predicate is contradictory. This fixes one of the major causes of it however.

Bug: https://github.com/google/xls/issues/2674
PiperOrigin-RevId: 786824998