Avoid calling SpecializeGivenPredicates in cases where the predicate is unsatisfiable.
NarrowingPass would combine context-qe and others in cases where the context-qe was unable to see that the predicate was unsatisfiable but other qes could. This could lead to the context-qe giving bounds that other qes see as impossible, potentially even to the point of being disjoint. Since the cases where this happen are unobservable anyway argurably this is all acceptable however in the case where there is no overlap between the two views of the world we would have check-failures. To avoid this simply have narrowing not attempt context narrowing in cases where a predicate is unsat; trusting other passes to remove the dead code eventually.

Fixes: https://github.com/google/xls/issues/2674
PiperOrigin-RevId: 786858190