[array_simp] Fix bug in array slice chain simplification
Rather than just checking if the end index runs off the end of the ultimate base array, we instead check if any intermediate slice can run off the end of the array it's slicing, and truncate the chain if it does.

This avoids complications that can occur when slicing "past the end" of an array, which will pad the array with extra copies of the last entry (to stay consistent with `array_index`).

Fixes google/xls#2692

PiperOrigin-RevId: 787263829