Add select lifting for binary operations.
Add support for selects that share the same operation in all of its cases,
e.g. `select(s, a + shared, b + shared) -> add(shared, select(s, a, b))`.

For now, because commutativity of the binary operation is not checked, we require that the shared operand is always on the LHS or always on the RHS, not sometimes one or the other.

PiperOrigin-RevId: 788151287