Fix miscompile with reassociation.
In some cases we could misoptimize math that includes subtraction of potentially INT_MIN values. It is difficult to observe this since it relies on extends that got folded in allowing visibility into overflow bits.

To fix it we pessimize in cases where int-min values might appear on the right hand side of subtractions by considering them overflowing if the add can overflow.

This was found with the ir fuzzer.

PiperOrigin-RevId: 788168684