Add helpers to write query-engine consistency fuzz tests.
This works by using the ir-fuzzer to generate a package and running a query engine over it. Callbacks are then performed with every node and the result of running the function on a set of inputs. The callback then can check that the nodes actual value is consistent with the query-engines abstract evaluation.

PiperOrigin-RevId: 789481984