Fix bitcount-qe bugs around 'negative' operation and add fuzz test.
Add a new fuzz test for bit-count query engine that checks that its abstract evaluation results are consistent with actual executions. This quickly found issues with the 'negative' operation which are also fixed in this cl.

It is likely that some combination of the irs compiled and the location in the pipeline meant this negative issue wasn't noticed.

PiperOrigin-RevId: 789483752