Add a "type_inference_v1" module attribute for forcing v1 once v2 is default.
This will be useful for the few tests that want to target v1 only, and also if some bug arises that motivates a temporary rollback to v1 to unblock development.

Although I am adding the attribute in some places where we know we want it, the attribute will be a no-op until TIv2 is actually on by default.

PiperOrigin-RevId: 789768055