Add --passes_profile flag
This flag records an (uncompressed) pprof file of the passes invoked during optimization. It records the number of times a pass was run and the duration of the pass run. This will allow us to identify outliers in pass performance and better track where the pass was invoked.

This adds a global state which keeps track of pass enter and leaves
as well as adding annotations. Annotations are label-like information.
A global atexit handler serializes to the pprof format.

PiperOrigin-RevId: 790855366