Fix bug with `S { data: other_s.data }` where `S` is parametric.
Over-aggressive filtering of the type annotations when inferring the implicit S parametrics was causing every annotation on `other_s` to be filtered out here, when in reality we want to filter none of those.

This change uses an annotation flag to more precisely target the filtering. It also improves the way flags are converted to string, which was getting unwieldy, and the change in filtering logic allows a good deal of code to be deleted.

PiperOrigin-RevId: 790978013