Added proc support to the non_synth_separation_pass.
We now separate out non-synthesizable nodes from procs. This is done by cloning the non-synth nodes of a proc into a function. This non-synth functions cannot contain the proc sends and receives, so instead the outputs of the sends and receives are passed to the function by parameters.

We still continue to remove non-synth nodes from the proc and add an invoke to the non-synth function.

PiperOrigin-RevId: 792246384