Fix check-fail in bytecode interpreter during range validation in TIv2.
This bug affected ranges with parametric-call endpoints in some rare situations.

While a populated `ParametricEnv` is not necessary to actually get parametric values in the scenario with the affected logic in validate_concrete_types.cc (they are in the `TypeInfo` object), bytecode interpreter wants the invocation data to be consistent.

PiperOrigin-RevId: 793639910