Use `final` where possible unless `override` is needed.
Readability: The `final` keyword indicates to the reader of
the code that an implementation is a leaf, improving readability;
it is always possible to make it `override` later.

Performance: This also can result in performance improvements as
the compiler can better see devirtualization opportunities.

This changes many places in leaf implementations to be
annotated `final` instead of `override`. Maybe it also helps future
code to adopt this good practice.