[xls] Fix order-of-argument-evaluation UB in PartialInformation constructors
Since some of the arguments were moved from on evaluation, our previous delegating-constructors design was unsafe. To avoid this, we switch from delegating to a common constructor to sharing a common (private) Initialize function after all members are initialized.

The new design is safe, since member initialization order is defined to be declaration order. This guarantees that the new value of `bit_count_` is evaluated before we can move from the provided `ternary` or `range`.

Fixes google/xls#2837

#gardening

PiperOrigin-RevId: 794574364