[opt] Prevent a self-replacement producing invalid IR in Reassociation
In an attempt to save computation, when replacing a node with a reassociation to `(x OP constant)`, we tried to force the new replacement node to use the same computed associative analysis as the node it was replacing (say, N0). However, if the constant was an identity, and N0 took another node N1 as a transitive operand, which in turn took `x` as a transitive operand, this could get us into trouble. When we later tried to reassociate N1, this would end up referencing `x` - but the analysis for `x` had been replaced with the analysis for N0, which could reference N1.

This produced circular references in the associative analyses, which could cause us to do invalid transformations on the IR, including attempting to replace a node with itself (which currently produces invalid IR).

While we're at it, we add a check that `ReplaceUsesWith` is not invoked with the node itself.

PiperOrigin-RevId: 794674946