Fixed a bug in a previous rolled back CL by avoiding the three way compare optimization if the compares are the same node.
Previous CL:
https://github.com/google/xls/commit/0c454ba54572acb86f9da274fd69a2f179d9a319

In the case that the first and second compare are the same node and that their lhs and rhs are the same operand, we want to avoid the optimization, otherwise changing one node will change both of them.

For example,
fn FuzzTest(p0: bits[1]) -> bits[1] {
  slt: bits[1] = slt(p0, p0)
  concat: bits[2] = concat(slt, slt, id=3)
  priority_sel: bits[1] = priority_sel(concat, cases=[slt, slt], default=p0)
}
would be changed to
fn FuzzTest(p0: bits[1] -> bits[1] {
  eq: bits[1] = eq(p0, p0)
  concat: bits[2] = concat(eq, eq)
  slt: bits[1] = slt(p0, p0)
  priority_sel: bits[1] = priority_sel(concat, cases=[p0, eq], default=eq)
}
where the slt node is swapped for the eq node in both cases.

PiperOrigin-RevId: 795539609