With the introduction of the param token-and-data producing op from the non-synth pass, avoid token dependencies with it.
The non_synth_separation_pass separates out non-synthasizable nodes into a differnt function from the original. It does this for both procs and functions. In the situation that a non-synth function is made from a proc such that a receive return value is passed as a parameter to the new function, this param node is considered a token-and-data producing op, but the token_dependency_pass only supports receive token-and-data producing ops. This change makes it so the params are ignored in this case.

PiperOrigin-RevId: 795579423