Fixes to reject indexing into a zero sized array.
Since the last element of the array does not exist we cannot use our typical
"clamp to last in bounds" out of bounds indexing behavior.

This uncovered that we were accidentally creating an array index to into a nil
tuple in the array simplifier, so we special case where there is only one
element to avoid creating this invalid IR. This made the fuzzing pass the new
verifier conditions.