[xls][mlir] Improve -optimize-spawns to take FifoConfigs into account
This pass was written before FifoConfigs existed. FifoConfigs currently exist on
SchanOps, but not on SprocOp arguments. -optimize-spawns works by eliminating a
redundant send/recv pair and merging the channels:

  chan a, b
  send ... to a
  ...
  recv %y from a
  send %y to b
  ...

  ~~>

  chan b
  send ... to b
  ...

Obviously we only see the initial `send` if `a` is an intermediate channel (an
SchanOp). Otherwise `a` would be an argument, and the sender is anonymous. The
converse is true for results and the `b` channel.

When `a` or `b` or both has a FifoConfig, we want to ensure sematics doesn't change
with this optimization. As we can't assign a FifoConfig to an argument or result, it
isn't correct to eliminate the local channel if the local channel has a FifoConfig.

However if both `a` and `b` are local channels, then we can fuse the two channels
together by merging their FifoConfigs.

PiperOrigin-RevId: 796372614