Avoid relying on all query engines having a full GetTernary implementation in UnionQueryEngine, Fix other QE Overrides. Implement GetTernary/Intervals for bit-count.
We found an odd fuzzer bug where due to the exact sequence of operations the bit-count query engine was able to determine the exact value of a node where other engines could not. This led to the narrowing pass to enter an unexpected condition. In order to fix this the bit-count query engine is updated to provide overrides for getting the known value, the query engine is updated to try to use the bit-count function for IsAllOnes/Zeros and the union-qe is updated to implement the known-value functions.

Together these ensure that the call to IsAllZeros in narrowing_pass.cc correctly returns True in the case where the bit-count-qe knows this but ternary does not.

To try to avoid issues like this occuring in the future this also implements the GetIntervals & GetTernary functions on bit-count query engine.

PiperOrigin-RevId: 797364267