Relax slice validation in TIv2 to be more on par with v1.
The extra range check is reduced to a warning.

This also stops generating negative width slice starts in the fuzzer, because v2 will still catch those in cases where v1 will not.

This change also includes a tweak to the conversion of TypeAnnotation to InferenceVariableKind because the fuzzer generates parametrics that are xN-typed.

With these changes, ast_generator_test and some integration tests will pass with v2 enabled by default.

PiperOrigin-RevId: 797418303