Add another narrow(Ternary) to pipeline soon after reassociation.
Otherwise other passes can interfere with narrowing seeing what is possible leading to infinite fixedpoint loops.

Also add a fuzz test which just checks that we can always optimize things the fuzzer creates in 1 million pass invocations. This limit was chosen arbitrarily so we might want to increase it or limit the fuzzer packages in some way.

Fixes: https://github.com/google/xls/issues/2890
PiperOrigin-RevId: 797776862