Partial unrolling optimization:
Try a trial unroll, with side effects masked, to see if a fixed number of iterations can be proven for the loop. If so, check if this is a multiple of the unrolling factor. If so, omit the conditions from the IR generated for the body of the pipelined loop. The condition is only evaluated to decide whether or not to exit the pipelined loop section, not for every partially unrolled iteration.

PiperOrigin-RevId: 797844232