Fix more tests for when TIv2 is turned on by default.
Note that the overshift in wide_ashr.x was only allowed by v1 due to its erroneous use of a truncated copy of the RHS when deciding if it's too large. In v2 we were attempting to do the same erroneous logic using a different function that errors. I have fixed both the v2 shift validation logic and the test.

PiperOrigin-RevId: 798208765