[opt] Add CriticalPathDelayAnalysis for arrival time estimates
This change introduces a new base class `LazyNodeData` from which `LazyNodeInfo` now inherits. `LazyNodeData` provides a general framework for caching node-specific data that depends on operand values. `LazyNodeInfo` specializes this for data stored in `LeafTypeTree`s.

A new analysis, `CriticalPathDelayAnalysis`, is added, which inherits from `LazyNodeData<int64_t>`. This analysis computes the critical path arrival time for each node in a function based on a provided `DelayEstimator`.

The `OptimizationContext` is updated to include a mechanism for caching and retrieving shared instances of `LazyNodeData` analyses.

We also set all optimizing rules to use the default delay model, while accepting user-specification as an override.

#geminiassisted

PiperOrigin-RevId: 800496993