Clarify logic around choosing the quickcheck evaluator.
Previously, the quickcheck evaluator piggy-backed on the comparison runner which meant if you ran the interpreter with --evaluator=jit --comparison=none you'd get a confusing message about not running the quickcheck tests because the jit is disabled. Now a separate options field holds the quickcheck runner. It uses the comparison one if comparison is enabled, otherwise uses the --evaluator one. Also, allow the ir-interpreter to be used for quickchecks but only if a special flag is passed in (--run_quickcheck_when_interpreting).