Get rid of ad hoc TIv1 use in VAST->DSLX for the purpose of dropping dead casts.
With this change, we optimistically insert casts, and then remove the dead ones after a real run of type inference.

After this change, there are still some uses of TIv1, which insert casts due to the differing type rules in DSLX vs. Verilog. The plan is to solve those with a type inference hook which will add more logic to the `DslxTypeFixer` added here.

PiperOrigin-RevId: 811349245