Remove capped `next_value_opt(3)` pass from optimization pipeline.
The `next_value_opt` pass is now run without a cap on its optimization level, as the default level is sufficient. The compound pass definition for `next_value_opt(3)` is also removed.

This has no affect on optimization decisions as nothing changes when opt level is higher than 3 (which is considered max level). It is mostly just a cleanup.

PiperOrigin-RevId: 817373790