Replace --emit_fail_as_assert flag with --emit_assert.
The name for this flag was quite bad as it implied that there may be some other cases where asserts could still be emitted. Instead change to the much more accurate "--emit_assert" flag to make clear that setting this to false causes **all** asserts to be suppressed.

The ir_converter_main binary will emit a warning if this flag is passed (ignoring the value) and the standard bazel rules for XLS will automatically swap --emit_fail_as_assert to --emit_assert (and print a warning).

PiperOrigin-RevId: 819958156