[opt] Split adders at points where carry propagation is impossible
This change introduces a new transformation in the strength reduction pass. If, based on ternary analysis, a bit position is found where a carry cannot propagate (i.e., both operands are known to be zero at that bit), the addition is split into two smaller additions concatenated together. This can reduce the critical path delay of the adder.

PiperOrigin-RevId: 820318959