Add support for #(...) style parameters to modules
This parameter form is used pretty universally and it's been supported since Verilog-2001 and always in SystemVerilog. In addition to being much more common than the old form, the new form enables use of parameters in input/output port definition not possible with parameter defined in the module body. Example:

module foo #(
  parameter int WIDTH = 8
) (
  input  logic [WIDTH-1:0] data_in,
  output logic [WIDTH-1:0] data_out
);
endmodule