Skip a failing test in OSS CI for JIT channel queues.
The `FixedValueGenerator` test in `LockLessJitChannelQueueTest` is failing in the Open Source CI environment. This change introduces a mechanism to skip this specific test when running in CI by defining a preprocessor macro `XLS_IS_CI_RUN` viathe  Bazel config option `--config=ci`. A virtual method is added to the test base to allow subclasses to indicate if they are in the OSS CI environment, enabling the conditional skip.

Issue: https://github.com/google/xls/issues/3223
PiperOrigin-RevId: 821868397