Add more type annotation flags to help both VAST->DSLX translation and error message tuning.
As of this change, the flags are not consumed yet.

A by-product is that this change corrects places in TIv2 where we were incorrectly using `TypeInferenceFlag::kNone` to mean "not one of the flags this place cares about." The corrections in tests to make a couple of types more longhand are symptoms of mis-application of kNone before, which was causing a user-specified type to be simplified when it should not be.

PiperOrigin-RevId: 822724466