Avoid using a JitRuntime over multiple Package*s in jit_channel_queue_test
This test was (correctly) transforming xls::Value into the jit representation used by our llvm jit. The JitRuntime which holds a map<xls::Type*, llvm::Type*> as a type cache was not invalidated or cleared for each Package*. This meant that if 2 tests happened to allocate xls::Type*s with the same pointer and they are different sizes then the tests could fail due to different pieces of code not agreeing on how much data needed to be copied. This caused the tests to read uninitialized memory.

To fix this rewrite the jit_channel_queue_test such that each run initializes its own JitRuntime* avoiding any sharing of caches.

Fixes: https://github.com/google/xls/issues/3223
PiperOrigin-RevId: 824535176