Disallow ram configurations with latency != 1 for now.
Ram latency needs to be rethought to support values != 1 (see google/xls#1053).  It is currently broken. The current behavior is to add a scheduling constraint of $LATENCY between request and response. This is not an unreasonable thing to do, but to get the correct behavior the receive on the response channel should wait $LATENCY cycles before capturing the response. It currently waits 1. Also, the response channel should have a $LATENCY-deep FIFO to catch responses in case of stalling (or a credit-like mechanism to stall the request side). Instead, the current behavior is a depth-1 FIFO that can overflow, resulting in lost data and potential deadlock. Until these issues are confronted, it seems best to require latency to be 1 as that is the only correctly supported configuration.

PiperOrigin-RevId: 825303986