When there is a TI unification error in a node context, include its span.
The motivating example is initialization of a struct instance field with an expr that has a conflicting declaration type. We would previously point to the two declaration spans in the error and not the place where you are trying to actually use the offending thing as the struct field value.

PiperOrigin-RevId: 828991144