AST-ify the attributes that appear before functions and types.
The hacky way this has been supported thus far has hit a wall, because we want to add `#[derive("Foo", "Bar", ...)]` and have that co-exist with other ones.

At this stage the parser converts the parsed attributes into the bespoke fields on Function, Proc, etc. that have always existed, and we make use of the new Attribute nodes only for ease of parsing and some minor formatting purposes. A later stage will use the uniform Attribute nodes more extensively.

PiperOrigin-RevId: 829552791