Set the parametric environment when creating a ParametricContext.
This change seems necessary to handle the cases where a proc alias to a parametric proc invokes a parametric function. The `SetParametricEnv()` call ensures that the newly created `ParametricContext` is properly associated with the provided `env`. If missing, it will try to get the env later, not find it, and have a default empty env. Subsequent parametric invocations will then wind up with an unexpected nullptr type_info.

Progress towards google/xls#890.

PiperOrigin-RevId: 831183135