Impose a path limit within BinaryDecisionDiagram.
In an effort to prevent BDD size explosion, modify BinaryDecisionDiagram's IfThenElse to saturate. BddNodeIndex(-1) is used to represent giving up and is accounted for by users of the BinaryDecisionDiagram, e.g SaturatingBddEvaluator converts this to `TooManyPaths`. This makes it possible for something like the VisibilityAnalysis to execute on tens of thousands of IR nodes in a reasonable time, properly giving up quickly on computing the BDD expression for simple parts that eventually add up to way too large a tree.

PiperOrigin-RevId: 831495704