Fix IR conversion of generified impl dispatch.
For example: `fn call_foo<T: type>(a: T) { a.foo(); }`

The reason this was working in the bytecode interpreter and not IR conversion was that bytecode interpreter had a longstanding hack around the use of TypeInfo for impl dispatch. Even that hack would stop working once we specialize trait derivations for the caller context.

This change makes TypeInfo::GetCallee and friends always return the right answer, even if you have a generified impl or we derive specialized trait implementations. Because that answer is not necessarily from the root TypeInfo now, there is some renaming.

PiperOrigin-RevId: 831932076