Introduce DataFlowLazyNodeInfo base class.
This class is meant to make it easy to implement lazy analyses that track leaf nodes, bits in this case as only Functions are supported, through compound operations (eg tuples) and invokes to their sources.

Note that if an operation is not explicitly supported, then the analysis will merge all of its operands. For example, for an array slice, the start index will be merged with the array being sliced. This can produce false positives, that is trace a node back to more sources than strictly necessary.

This was implemented in order to enable XLS[cc] to track direct-in values through function slices, from input parameters to continuation outputs. This allows direct-ins not to be placed into StateElements.

Also remove currently unneeded optional return to LazyNodeData::GetInfo().

PiperOrigin-RevId: 832364864