Run PartialInformation reconciliation to fixed point
The number of times information must be synthesized between the ternary and range representation is variable based on the exact contents of the information. Simply iterate on the reconciliation until a fixed point is achieved to avoid missing anything.

Also update ternary_ops to track this information during computation of the union since it avoids a second scan of the vectors.

PiperOrigin-RevId: 832415491