Rewrite GatherBits(Node*, LTT<Bits> mask).
This function was simply written incorrectly for a few reasons. (1) It used a Span as a map key when the memory the span points to was not constant and (2) It incorrectly iterated through the LTT.

To fix this the function was rewritten using the ToTreeOfNodes function enabling the whole thing to be much simpler.

PiperOrigin-RevId: 838847612