[codegen 1.5] Implement FlowControlInsertionPass
This pass modifies ScheduledBlocks to insert the necessary logic for pipeline flow control, replacing the placeholder `inputs_valid` and `outputs_ready` nodes with actual logic involving registers and combinations of stage flow-control signals.

Unlike most of our other passes, this one is *not* idempotent, as there's no simple way for it to tell whether these nodes are placeholders or have already been replaced. However, it shouldn't do any harm to run it a finite number of times; just be aware that it should never been run to fixed point, as it will never converge.

PiperOrigin-RevId: 839251688