Ensure next_values(StateRead*) returns an empty set in the rare case where no next's are associated with a read
This is a very rare situation but it can happen during next-value optimization passes and a few other places. To simplify the implementation of these passes return an empty set in these cases. Also check that the side-table maps used to determine this are valid in this case to ensure we don't ignore corrupted procs.

PiperOrigin-RevId: 840426382