Add ScheduledBlockConversionPass in codegen v1.5.
This converts a ScheduledFunction or ScheduledProc into a ScheduledBlock that contains a skeletal source entity and possibly proc-specific node types. This way we can incrementally lower source constructs into proper blocks.

To allow proc I/O against source entity channels in a partially lowered block, there are some changes in the internals of various IR functionality to remove assumptions that I/O is in a proc.

PiperOrigin-RevId: 841974389