Get traits and generics working with proc-scoped channels.
The core of the fix is using the right TypeInfo for the contents of to_bits() or a generic function. The prior logic for determining the effective TypeInfo for an Invocation in GetConversionRecords was inadequate for these cases. With this change, we just have TypeInfo provide a list of all the concrete scenarios for an Invocation node.

PiperOrigin-RevId: 842256494