Fix timeouts in GetConversionRecords.
One problem was that the scheme for discovery of foreign functions has for a while inadvertently relied on a subordinate ConversionOrderVisitor having the wrong associated module. The wrong module argument caused a visitor for module foo to keep thinking helper functions within foo needed foreign treatment with respect to foo.

A second problem was that, when ConvertInvocation was recently fixed to correctly find all concrete calls per invocation, the pre-existing potential to re-encounter the same Invocation multiple times became too expensive, and the expense was compounded by the other bug. This change short-circuits re-processing of an Invocation node.

PiperOrigin-RevId: 843256790