Rewrite CSE to create nicer verilog
CSE would select the new 'real' node in a pretty random seeming manner (related to the exact node-ids everything is assigned in the end). With things like non-synth separation this can lead to some nodes getting unexpected names. This changes the CSE to prioritize making the real node the one with the shortest name.

This also rewrites how CSE is performed slightly in a way that should make it converge more quickly by handling commutative operations slightly better

PiperOrigin-RevId: 843372059