Fix the scheduled IR dumping logic to avoid out-of-oder stageless nodes.
The problem with the prior approach is that it relies on a non-transitive analysis of what stage it can put each stageless node adjacent to. This reordering can cause a series of stageless nodes to get un-toposorted with respect to each other. We would see this start to become a problem with the _full bit logic in proc state lowering.

PiperOrigin-RevId: 844775185